+++ src/main/java/com/service/BOOKJEOK/security/dto/CustomOauth2UserDetails.java
@@ -1,0 +1,71 @@
+package com.service.BOOKJEOK.security.dto;
+
+import com.service.BOOKJEOK.domain.User;
+import com.service.BOOKJEOK.security.oauth.provider.OAuthAttributes;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Map;
+
+@Getter
+@RequiredArgsConstructor
+public class CustomOauth2UserDetails implements OAuth2User, UserDetails {
+
+    private final User user;
+    private final OAuthAttributes attributes;
+
+    @Override
+    public String getPassword() {
+        return "null";
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getName();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return attributes.getAttributes();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        String role = user.getRole().value();
+        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(role);
+        authorities.add(grantedAuthority);
+        return authorities;
+    }
+
+    @Override
+    public String getName() {
+        return user.getProviderId();
+    }
+}
