+++ src/main/java/com/service/BOOKJEOK/security/dto/PrincipalUserDetails.java
@@ -1,0 +1,55 @@
+package com.service.BOOKJEOK.security.dto;
+
+import com.service.BOOKJEOK.domain.User;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+@RequiredArgsConstructor
+public class PrincipalUserDetails implements UserDetails {
+
+    private final User user;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        String role = user.getRole().value();
+        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(role);
+        authorities.add(grantedAuthority);
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return "null";
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getName();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
